Fix fsa computation in feltor
